Implemented 'top score' route

+++ project2/src/main/java/csc312/countdowns/CountdownCallback.java
@@ -5,7 +5,7 @@ public interface CountdownCallback
     
     void onBegin();
     
-    void onTick(long remainingTime);
+    void onTick(int remainingTime);
     
     void onComplete();
